Fix Serde Test
The bug's pretty simple java.awt.Color is a class so we can't use simple
`!=` syntax to compare it, because that'll use reference equality only.

Digging through I ended up adding ToString to all the various classes
involved in the LootFilter so that the test produces a more readable
error comparison.